feat: Log in / Create account opens as a Google-auth popup
Soft in-app navigations to /login are intercepted (app/@authmodal) and
render the Google-auth explainer as a modal over the launching page —
every existing <Link href="/login"> CTA gets the popup with no changes.
Hard navigations (invite emails, the callback's auth-failure redirect,
refreshes, the middleware's signed-out redirect) still get the full page.

The explainer card is extracted to login-card.tsx — one copy source for
both hosts, identical auth semantics (signInWithOAuth, invite cookie).
The modal shell reuses the .gate-modal/.gate-sheet vocabulary; X, Esc,
backdrop, and browser-back all return to the launching page.
