+++ app/@authmodal/default.tsx
@@ -1,0 +1,4 @@
+// The @authmodal slot is empty everywhere except an intercepted /login.
+export default function Default() {
+  return null;
+}
